Change basic algorithm
Before, this was finding all words and then subsequently scanning the
text for each word for snippeting purposes.  In practice, that meant
it was O(N^2) since it had to scan the whole text for each word.
Instead, don't use regexes for finding words (to select them more
carefully, especially around apostrophes) and snippet them immediately
since we're already at the proper index.

Much faster and better results to boot.